refactor: extract Diagnostic namespace into lsp/diagnostic.ts + self-reexport
The Diagnostic namespace was the only `export namespace` in lsp/lsp.ts
but lived alongside flat top-level exports (Event, Range, Symbol, etc.)
for the main LSP service. A mechanical unwrap would have projected the
whole file as Diagnostic (LSP.Diagnostic.Range would exist) which is
cosmetically wrong.

Instead, extract Diagnostic to its own file lsp/diagnostic.ts with a
self-reexport (`export * as Diagnostic from "./diagnostic"`), and
re-export it from lsp.ts to preserve the LSP.Diagnostic.pretty /
LSP.Diagnostic.report consumer ergonomics.

Public API unchanged. Consumer imports unchanged.

Verified: tsgo --noEmit 0 errors, SDK generate no drift,
test/lsp/ 18/18 pass, tool tests using LSP.Diagnostic (write, edit,
apply_patch) 66/66 pass.
